+++ NEWS.md
@@ -1,8 +1,18 @@
 ### Version 2.4.0
 
-#### Fixes:
+#### Improvements:
 
    - updates int the UI design
+
+### Version 2.3.5
+
+#### Fixes:
+
+   - fixes delays in the process loop and improve the process loop
+
+#### Other
+
+   - remove experimental file system support
 
 ### Version 2.3.4